Fix paths after module reorg
