+++ src/core/mod.rs
@@ -1,0 +1,5 @@
+pub mod message;
+pub mod config;
+
+pub use message::{Message, MessageVecExt};
+pub use config::GenerationConfig;
