+++ src/loaders.rs
@@ -1,6 +1,7 @@
 use hf_hub::api::tokio::Api as HfApi;
 use std::path::PathBuf;
 use tokenizers::Tokenizer;
+use crate::core::GenerationConfig;
 
 #[derive(Debug, Clone)]
 pub struct HfLoader {
@@ -70,18 +71,6 @@ impl TokenizerLoader {
         Ok(tokenizer)
     }
 }
-
-#[derive(Clone)]
-pub struct GenerationConfig {
-    pub temperature: Option<f64>,
-    pub top_p: Option<f64>,
-    pub top_k: Option<u64>,
-    pub min_p: Option<f64>,
-    pub repeat_penalty: Option<f32>,
-    pub repeat_last_n: Option<usize>,
-    pub eos_token_ids: Vec<u64>,
-}
-
 pub struct GenerationConfigLoader {
     pub generation_config_file_loader: HfLoader,
 }
