+++ src/models/components/attention.rs
@@ -1,0 +1,15 @@
+use candle_core::{Result, Tensor};
+
+/// Repeats a key or value tensor for grouped query attention.
+/// The input tensor should have a shape `(batch, num_kv_heads, seq_len, head_dim)`.
+pub fn repeat_kv(xs: Tensor, n_rep: usize) -> Result<Tensor> {
+    if n_rep == 1 {
+        Ok(xs)
+    } else {
+        let (b_sz, n_kv_head, seq_len, head_dim) = xs.dims4()?;
+        // Using cat is faster than a broadcast as it avoids going through a potentially
+        // strided copy.
+        // https://github.com/huggingface/candle/pull/2043
+        Tensor::cat(&vec![&xs; n_rep], 2)?.reshape((b_sz, n_kv_head * n_rep, seq_len, head_dim))
+    }
+}
