+++ src/models/implementations/mod.rs
@@ -1,0 +1,3 @@
+pub mod modernbert;
+pub mod gemma3;
+pub mod qwen3;
