+++ src/models/mod.rs
@@ -1,159 +1,20 @@
 pub mod generation;
-pub mod quantized_nn;
 
-pub mod modernbert;
-pub mod quantized_gemma3;
-pub mod quantized_qwen3;
+pub mod components {
+    pub mod layers;
+    pub mod attention;
+    pub mod quantization;
 
-pub use quantized_nn::RmsNorm;
-
-use candle_core::{Result, Tensor};
-use candle_nn::Module;
-
-/// Repeats a key or value tensor for grouped query attention
-/// The input tensor should have a shape `(batch, num_kv_heads, seq_len, head_dim)`,
-pub fn repeat_kv(xs: Tensor, n_rep: usize) -> Result<Tensor> {
-    if n_rep == 1 {
-        Ok(xs)
-    } else {
-        let (b_sz, n_kv_head, seq_len, head_dim) = xs.dims4()?;
-        // Using cat is faster than a broadcast as it avoids going through a potentially
-        // strided copy.
-        // https://github.com/huggingface/candle/pull/2043
-        Tensor::cat(&vec![&xs; n_rep], 2)?.reshape((b_sz, n_kv_head * n_rep, seq_len, head_dim))
-    }
+    pub use layers::RmsNorm;
+    pub use quantization::{QMatMul, VarBuilder};
+    pub use attention::repeat_kv;
 }
 
-// QMatMul wrapper adding some tracing.
-#[derive(Clone)]
-pub struct QMatMul {
-    inner: candle_core::quantized::QMatMul,
-    span: tracing::Span,
+pub mod implementations {
+    pub mod modernbert;
+    pub mod gemma3;
+    pub mod qwen3;
 }
 
-impl QMatMul {
-    pub fn new(out_dim: usize, in_dim: usize, vb: VarBuilder) -> Result<Self> {
-        let ws = vb.get((in_dim, out_dim), "weight")?;
-        let inner = candle_core::quantized::QMatMul::from_arc(ws)?;
-        let span = tracing::span!(tracing::Level::TRACE, "qmatmul");
-        Ok(Self { inner, span })
-    }
-
-    pub fn from_weights(ws: std::sync::Arc<candle_core::quantized::QTensor>) -> Result<Self> {
-        let inner = candle_core::quantized::QMatMul::from_arc(ws)?;
-        let span = tracing::span!(tracing::Level::TRACE, "qmatmul");
-        Ok(Self { inner, span })
-    }
-}
-
-impl Module for QMatMul {
-    fn forward(&self, xs: &Tensor) -> Result<Tensor> {
-        let _enter = self.span.enter();
-        self.inner.forward(xs)
-    }
-}
-
-impl std::fmt::Debug for QMatMul {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "QMatMul")
-    }
-}
-
-use candle_core::quantized::QTensor;
-use candle_core::{Device, Shape};
-use std::sync::Arc;
-
-// VarBuilder specialized for QTensors
-#[derive(Clone)]
-pub struct VarBuilder {
-    data: Arc<std::collections::HashMap<String, Arc<QTensor>>>,
-    path: Vec<String>,
-    device: Device,
-}
-
-impl VarBuilder {
-    pub fn from_gguf<P: AsRef<std::path::Path>>(p: P, device: &Device) -> Result<Self> {
-        let mut file = std::fs::File::open(p)?;
-        let content = candle_core::quantized::gguf_file::Content::read(&mut file)?;
-        let mut data = std::collections::HashMap::new();
-        for tensor_name in content.tensor_infos.keys() {
-            let tensor = content.tensor(&mut file, tensor_name, device)?;
-            data.insert(tensor_name.to_string(), Arc::new(tensor));
-        }
-        Ok(Self {
-            data: Arc::new(data),
-            path: Vec::new(),
-            device: device.clone(),
-        })
-    }
-
-    pub fn from_gguf_buffer(buffer: &[u8], device: &Device) -> Result<Self> {
-        let mut cursor = std::io::Cursor::new(buffer);
-        let content = candle_core::quantized::gguf_file::Content::read(&mut cursor)?;
-        let mut data = std::collections::HashMap::new();
-        for tensor_name in content.tensor_infos.keys() {
-            let tensor = content.tensor(&mut cursor, tensor_name, device)?;
-            data.insert(tensor_name.to_string(), Arc::new(tensor));
-        }
-        Ok(Self {
-            data: Arc::new(data),
-            path: Vec::new(),
-            device: device.clone(),
-        })
-    }
-
-    pub fn pp<S: ToString>(&self, s: S) -> Self {
-        let mut path = self.path.clone();
-        path.push(s.to_string());
-        Self {
-            data: self.data.clone(),
-            path,
-            device: self.device.clone(),
-        }
-    }
-
-    fn path(&self, tensor_name: &str) -> String {
-        if self.path.is_empty() {
-            tensor_name.to_string()
-        } else {
-            [&self.path.join("."), tensor_name].join(".")
-        }
-    }
-
-    pub fn get<S: Into<Shape>>(&self, s: S, name: &str) -> Result<Arc<QTensor>> {
-        let path = self.path(name);
-        match self.data.get(&path) {
-            None => {
-                candle_core::bail!("cannot find tensor {path}")
-            }
-            Some(qtensor) => {
-                let shape = s.into();
-                if qtensor.shape() != &shape {
-                    candle_core::bail!(
-                        "shape mismatch for {name}, got {:?}, expected {shape:?}",
-                        qtensor.shape()
-                    )
-                }
-                Ok(qtensor.clone())
-            }
-        }
-    }
-
-    pub fn get_no_shape(&self, name: &str) -> Result<Arc<QTensor>> {
-        let path = self.path(name);
-        match self.data.get(&path) {
-            None => {
-                candle_core::bail!("cannot find tensor {name}")
-            }
-            Some(qtensor) => Ok(qtensor.clone()),
-        }
-    }
-
-    pub fn device(&self) -> &Device {
-        &self.device
-    }
-
-    pub fn contains_key(&self, key: &str) -> bool {
-        self.data.contains_key(key)
-    }
-}
+pub use components::{repeat_kv, QMatMul, RmsNorm, VarBuilder};
+pub use implementations::{gemma3 as quantized_gemma3, modernbert, qwen3 as quantized_qwen3};
