+++ src/pipelines/mod.rs
@@ -1,9 +1,8 @@
-pub(crate) mod utils;
-
+// Re-export the global cache from the utils module
 #[doc(hidden)]
-pub use utils::model_cache::global_cache;
+pub use crate::utils::cache::global_cache;
 
 pub mod fill_mask_pipeline;
 pub mod sentiment_analysis_pipeline;
-pub mod text_generation_pipeline;
+pub mod text_generation;
 pub mod zero_shot_classification_pipeline;
