+++ src/pipelines/text_generation/completion_stream.rs
