+++ src/pipelines/text_generation/event_stream.rs
