+++ src/pipelines/text_generation/tool_error.rs
