+++ src/pipelines/text_generation/xml_parser.rs
