+++ src/utils/mod.rs
@@ -1,6 +1,6 @@
 use candle_core::{CudaDevice, Device};
 
-pub mod model_cache;
+pub mod cache;
 
 /// Loads a device to be used for the model.
 /// If `index` is `Some(i)` it will attempt to load the specified CUDA device.
